requirements.txt
# ==================== Agent Backend API 依赖包 ====================
# 基于项目代码分析生成的完整依赖列表

# ==================== Web框架核心 ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ==================== HTTP客户端 ====================
requests>=2.31.0
httpx>=0.25.0  # 异步HTTP客户端，用于更好的性能

# ==================== 数据库相关 ====================
sqlalchemy>=2.0.0
#sqlite3  # Python内置，但明确列出

# ==================== AI/ML相关 ====================
# Transformers库用于情感分析模型
transformers>=4.52.1
#torch
torchaudio>=2.8.0
tokenizers>=0.21.0
safetensors>=0.5.2

# ==================== 音频处理 ====================
# IndexTTS相关依赖
librosa>=0.10.2
soundfile>=0.12.1
numpy>=1.26.2
scipy>=1.11.0

# ==================== 文本处理 ====================
jieba>=0.42.1  # 中文分词
cn2an>=0.5.22  # 中文数字转换
g2p-en>=2.1.0  # 英文音素转换
wetext>=0.0.9  # 文本处理工具

# ==================== 配置和工具 ====================
omegaconf>=2.3.0  # 配置管理
python-multipart>=0.0.6  # 文件上传支持
python-json-logger>=2.0.0  # JSON日志格式

# ==================== 开发和调试 ====================
python-dotenv>=1.0.0  # 环境变量管理
loguru>=0.7.0  # 更好的日志库

# ==================== 类型提示支持 ====================
typing-extensions>=4.8.0

# ==================== 数据处理 ====================
pandas>=2.3.2  # 数据处理
munch>=4.0.0  # 字典访问工具
pytz>=2023.3  # 时区处理

# ==================== 图像处理（如果需要） ====================
opencv-python>=4.9.0
Pillow>=10.0.0

# ==================== 数学和科学计算 ====================
matplotlib>=3.8.2
tqdm>=4.67.1  # 进度条

# ==================== 加速和优化 ====================
accelerate>=1.8.1  # 模型加速
einops>=0.8.1  # 张量操作

# ==================== 其他工具 ====================
pathlib2>=2.3.7; python_version < "3.4"  # 路径处理兼容性
json5>=0.10.0  # JSON5支持
textstat>=0.7.10  # 文本统计

# ==================== 可选依赖（根据需要安装） ====================
# 如果需要WebSocket支持
 websockets>=11.0.0

# 如果需要Redis缓存
 redis>=5.0.0

# 如果需要PostgreSQL支持
# psycopg2-binary>=2.9.0

# 如果需要MySQL支持
# PyMySQL>=1.1.0

# 如果需要更好的异步支持
# asyncio-mqtt>=0.16.0

# 如果需要任务队列
# celery>=5.3.0

# 如果需要监控
# prometheus-client>=0.19.0

# ==================== 开发和测试依赖 ====================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# ==================== 系统要求 ====================
# Python >= 3.10
# CUDA支持（可选，用于GPU加速）
# FFmpeg（用于音频处理）

